feat: run tests in browser by demand
