+++ lib/test/karma/bootstrap.js
@@ -1,0 +1,23 @@
+const { expect, use } = require('chai');
+const dirtyChai = require('dirty-chai');
+const chaiAsPromised = require('chai-as-promised');
+const sinon = require('sinon');
+const sinonChai = require('sinon-chai');
+
+use(chaiAsPromised);
+use(dirtyChai);
+use(sinonChai);
+
+beforeEach(function beforeEach() {
+  if (!this.sinon) {
+    this.sinon = sinon.createSandbox();
+  } else {
+    this.sinon.restore();
+  }
+});
+
+afterEach(function afterEach() {
+  this.sinon.restore();
+});
+
+global.expect = expect;
