+++ package.json
@@ -7,7 +7,8 @@
     "test": "mocha './test/**/*.spec.js'",
     "lint": "eslint .",
     "test:e2e": "NODE_ENV=test mocha 'test/e2e/**/*.spec.js'",
-    "test:functional": "NODE_ENV=test mocha 'test/functional/**/*.spec.js'"
+    "test:functional": "NODE_ENV=test mocha 'test/functional/**/*.spec.js'",
+    "test:browsers": "karma start ./karma.conf.js --single-run"
   },
   "repository": {
     "type": "git",
@@ -21,23 +22,39 @@
   "dependencies": {
     "@dashevo/dashcore-lib": "~0.19.26",
     "@dashevo/dpns-contract": "^0.4.0",
-    "@dashevo/feature-flags-contract": "^0.2.1",
     "@dashevo/dpp": "~0.21.0-dev.7",
+    "@dashevo/feature-flags-contract": "^0.2.1",
     "@dashevo/grpc-common": "~0.5.4",
     "@dashevo/merk": "github:dashevo/node-merk",
     "@dashevo/wallet-lib": "~7.21.0-dev.8",
+    "assert-browserify": "^2.0.0",
     "blake3": "^2.1.4",
+    "buffer": "^6.0.3",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
+    "crypto-browserify": "^3.12.0",
     "dash": "~3.21.0-dev.8",
     "dirty-chai": "^2.0.1",
     "dotenv-safe": "^8.2.0",
     "github-api": "^3.3.0",
     "js-merkle": "^0.1.5",
+    "karma": "^6.3.4",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-firefox-launcher": "^2.1.1",
+    "karma-mocha": "^2.0.1",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-webpack": "^5.0.0",
     "mocha": "^7.2.0",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "^1.0.1",
+    "process": "^0.11.10",
     "semver": "^7.3.2",
     "sinon": "^9.2.0",
-    "sinon-chai": "^3.5.0"
+    "sinon-chai": "^3.5.0",
+    "stream-browserify": "^3.0.0",
+    "url": "^0.11.0",
+    "webpack": "^5.55.1"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
